+++ src/primitives/arc/linear_equation.rs
@@ -20,10 +20,17 @@ pub struct LinearEquation {
 impl LinearEquation {
     /// Create a new linear equation based on one point and one angle
     pub fn from_point_angle(point: Point, angle: Angle) -> Self {
-        let (a, b) = match angle.tan() {
-            None => (Real::from(1.0), Real::from(0.0)),
-            Some(a) => (-a, Real::from(-1.0)),
+        // FIXME: angle.tan() for 180.0 degrees isn't exactly 0 which causes problems when drawing
+        //        a single quadrant. Is there a better solution to fix this?
+        let (a, b) = if angle == Angle::from_degrees(180.0) {
+            (Real::from(0.0), Real::from(-1.0))
+        } else {
+            match angle.tan() {
+                None => (Real::from(1.0), Real::from(0.0)),
+                Some(a) => (-a, Real::from(-1.0)),
+            }
         };
+
         let c = -(a * point.x.into() + b * point.y.into());
         LinearEquation { a, b, c }
     }
@@ -37,12 +44,15 @@ impl LinearEquation {
         }
     }
 
-    /// Check on which side of the line a point is
-    pub fn side(&self, point: Point) -> LineSide {
-        if self.a * point.x.into() + self.b * point.y.into() + self.c < Real::from(0.0) {
-            LineSide::Below
-        } else {
-            LineSide::Above
+    /// Checks if a point is on the given side of the line.
+    ///
+    /// Always returns `true` if the point is on the line.
+    pub fn check_side(&self, point: Point, side: LineSide) -> bool {
+        let t = self.a * point.x.into() + self.b * point.y.into() + self.c;
+
+        match side {
+            LineSide::Below => t <= Real::from(0.0),
+            LineSide::Above => t >= Real::from(0.0),
         }
     }
 }